Refactor imports in test files to streamline type imports from "@/types" instead of "@/types/api". This change enhances consistency across the codebase.
