+++ src/app/api/v1/data-connections/[data_connection_id]/data-connections.data_connection_id.test.ts
@@ -18,7 +18,7 @@ import {
   UserSession,
   DataConnection,
   RepositoryDataMode,
-} from "@/types/api";
+} from "@/types";
 import { AccountType } from "@/types/account";
 
 jest.mock("@/api/utils", () => ({
