+++ src/app/api/v1/memberships/[membership_id]/revoke.test.ts
@@ -11,12 +11,7 @@ import {
   BadRequestError,
 } from "@/lib/api/errors";
 import { MockNextApiResponse, jsonBody } from "@/api/utils/mock";
-import {
-  Membership,
-  MembershipRole,
-  MembershipState,
-  Actions,
-} from "@/types/api";
+import { Membership, MembershipRole, MembershipState, Actions } from "@/types";
 
 jest.mock("@/api/utils", () => ({
   getServerSession: jest.fn(),
