+++ src/app/api/v1/repositories/[account_id]/[repository_id]/api-keys/api-keys.test.ts
@@ -21,7 +21,7 @@ import {
   RepositoryDataMode,
   RepositoryFeatured,
   Actions,
-} from "@/types/api";
+} from "@/types";
 import { AccountType } from "@/types/account";
 
 jest.mock("@/api/utils", () => ({
