+++ src/app/api/v1/repositories/[account_id]/[repository_id]/members/members.test.ts
@@ -25,7 +25,7 @@ import {
   RepositoryDataMode,
   RepositoryFeatured,
   Actions,
-} from "@/types/api";
+} from "@/types";
 import { AccountType } from "@/types/account";
 import { Account } from "@/types/account";
