+++ src/lib/clients/database/api-keys.ts
@@ -1,4 +1,4 @@
-import { APIKey } from "@/types/api";
+import { APIKey } from "@/types";
 import { PutItemCommand } from "@aws-sdk/client-dynamodb";
 import { QueryCommand } from "@aws-sdk/lib-dynamodb";
 import { marshall } from "@aws-sdk/util-dynamodb";
