+++ src/lib/clients/database/memberships.ts
@@ -1,4 +1,4 @@
-import { Membership } from "@/types/api";
+import { Membership } from "@/types";
 import { PutItemCommand } from "@aws-sdk/client-dynamodb";
 import { QueryCommand } from "@aws-sdk/lib-dynamodb";
 import { marshall } from "@aws-sdk/util-dynamodb";
