+++ src/types/index.ts
@@ -1,19 +1,11 @@
 export * from "./shared";
-
 export * from "./account";
-
 export * from "./api-key";
-
 export * from "./data-connection";
-
 export * from "./membership";
-
 export * from "./product";
 export * from "./product_v2";
 export * from "./product_object";
-
 export * from "./session";
-
 export * from "./storage";
-
 export * from "./form";
